mds: __u32, not long

+++ src/mds/Migrator.cc
@@ -959,7 +959,7 @@ int Migrator::encode_export_dir(bufferlist& exportbl,
   ::encode(df, exportbl);
   dir->encode_export(exportbl);
   
-  long nden = dir->items.size();
+  __u32 nden = dir->items.size();
   ::encode(nden, exportbl);
   
   // dentries
@@ -2108,7 +2108,7 @@ int Migrator::decode_import_dir(bufferlist::iterator& blp,
   dout(15) << "doing contents" << dendl;
   
   // contents
-  long nden;
+  __u32 nden;
   ::decode(nden, blp);
   
   for (; nden>0; nden--) {
